+++ tests.html
@@ -24,18 +24,20 @@
         along with Kiwix (file LICENSE-GPLv3.txt).  If not, see <http://www.gnu.org/licenses/>
         -->
 
-        <link rel="stylesheet" href="tests/qunit-2.3.2.css" />
-        <script src="tests/qunit-2.3.2.js"></script>
+        <link rel="stylesheet" href="tests/qunit-2.11.3.css" />
+        <script src="tests/qunit-2.11.3.js"></script>
+        
+        
 
         <!-- Using require.js, a module system for javascript, include the
              js files. This loads "main.js", which in turn can load other
              files, all handled by require.js:
              http://requirejs.org/docs/api.html#jsfiles -->
-        <script type="text/javascript" src="www/js/lib/xzdec.js"></script>
-        <script type="text/javascript"
-                data-main="tests/init.js"
-        src="www/js/lib/require.js"></script>
-
+             <script type="text/javascript" src="www/js/lib/xzdec.js"></script>
+             <script type="text/javascript"
+                     data-main="tests/init.js"
+             src="www/js/lib/require.js"></script>
+     
 
     </head>
     <body>
@@ -43,5 +45,7 @@
         Another option is to force your browser to accept that (but you'll open a security breach) : on Chrome, you can start it with --allow-file-access-from-files command-line argument; on Firefox, you can set privacy.file_unique_origin to false in about:config
         <div id="qunit"></div>
         <div id="qunit-fixture"></div>
+       
+
     </body>
 </html>
