+++ tests/qunit-2.11.3.css
@@ -1,15 +1,26 @@
 /*!
- * QUnit 2.3.2
+ * QUnit 2.11.3
  * https://qunitjs.com/
  *
- * Copyright jQuery Foundation and other contributors
+ * Copyright OpenJS Foundation and other contributors
  * Released under the MIT license
  * https://jquery.org/license
  *
- * Date: 2017-04-18T02:19Z
+ * Date: 2020-10-05T01:34Z
  */
 
 /** Font Family and Sizes */
+
+/* Style our buttons in a simple way, uninfluenced by the styles
+   the tested app might load. Don't affect buttons in #qunit-fixture!
+   https://github.com/qunitjs/qunit/pull/1395
+   https://github.com/qunitjs/qunit/issues/1437 */
+#qunit-testrunner-toolbar button,
+#qunit-testresult button {
+	font-size: initial;
+	border: initial;
+	background-color: buttonface;
+}
 
 #qunit-tests, #qunit-header, #qunit-banner, #qunit-testrunner-toolbar, #qunit-filteredTest, #qunit-userAgent, #qunit-testresult {
 	font-family: "Helvetica Neue Light", "HelveticaNeue-Light", "Helvetica Neue", Calibri, Helvetica, Arial, sans-serif;
@@ -98,6 +109,10 @@
 	height: 1.6em;
 }
 
+#qunit-toolbar-filters {
+	float: right;
+}
+
 .qunit-url-config,
 .qunit-filter,
 #qunit-modulefilter {
@@ -107,7 +122,6 @@
 
 .qunit-filter,
 #qunit-modulefilter {
-	float: right;
 	position: relative;
 	margin-left: 1em;
 }
@@ -118,7 +132,7 @@
 
 #qunit-modulefilter-search {
 	box-sizing: border-box;
-	width: 400px;
+	min-width: 400px;
 }
 
 #qunit-modulefilter-search-container:after {
@@ -131,7 +145,7 @@
 #qunit-modulefilter-dropdown {
 	/* align with #qunit-modulefilter-search */
 	box-sizing: border-box;
-	width: 400px;
+	min-width: 400px;
 	position: absolute;
 	right: 0;
 	top: 50%;
@@ -204,6 +218,7 @@
 #qunit-modulefilter-dropdown-list .clickable {
 	display: block;
 	padding-left: 0.15em;
+	padding-right: 0.5em;
 }
